requirements/dev.txt
-r production.txt
-r lint.txt
-r test.txt

django-debug-toolbar

ipython
